# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "نبض الحياة API",
        "description": "API للتطبيق الخيري نبض الحياة",
        "contact": {
            "email": "info@nabdhayah.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/campaigns": {
            "get": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "الحصول على قائمة الحملات",
                "operationId": "00dd616cf477e3ccfcc20fd15a288de0",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "رقم الصفحة",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "عدد العناصر في الصفحة",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "البحث في العنوان والوصف",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "مساعدة"
                        }
                    },
                    {
                        "name": "priority",
                        "in": "query",
                        "description": "الحملات ذات الأولوية فقط",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "example": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "قائمة الحملات",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CampaignResource"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 50
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "إنشاء حملة جديدة",
                "operationId": "1a576fa89ebd794d75e4d0a5d9e0420d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "target_amount"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "مساعدة الأسر المحتاجة"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "حملة لمساعدة الأسر المحتاجة في رمضان"
                                    },
                                    "target_amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 10000
                                    },
                                    "images[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "تم إنشاء الحملة بنجاح",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "تم إنشاء الحملة بنجاح"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/CampaignResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "خطأ في البيانات",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "خطأ في البيانات"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/campaigns/{id}": {
            "get": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "الحصول على تفاصيل حملة",
                "operationId": "2d3f5dbf8fbbd91d2740579d53633602",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "معرف الحملة",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "تفاصيل الحملة",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/CampaignResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "الحملة غير موجودة",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "الحملة غير موجودة"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/campaigns/{id}/favorite": {
            "post": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "إضافة/إزالة حملة من المفضلة",
                "operationId": "0047d2b01a640910d23610ea549ff3cb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "معرف الحملة",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "تم تحديث المفضلة بنجاح",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "تم إضافة الحملة للمفضلة"
                                        },
                                        "is_favorited": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/my-campaigns": {
            "get": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "الحصول على حملات المستخدم",
                "operationId": "e9fc3c3a74951acd07ea2d52c5084c29",
                "responses": {
                    "200": {
                        "description": "حملات المستخدم",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CampaignResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/favorites": {
            "get": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "الحصول على الحملات المفضلة",
                "operationId": "7ecb1a367df1646a390a18c5ce51d84c",
                "responses": {
                    "200": {
                        "description": "الحملات المفضلة",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CampaignResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "CampaignResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "مساعدة الأسر المحتاجة"
                    },
                    "description": {
                        "type": "string",
                        "example": "حملة لمساعدة الأسر المحتاجة في رمضان"
                    },
                    "target_amount": {
                        "type": "number",
                        "format": "float",
                        "example": 10000
                    },
                    "current_amount": {
                        "type": "number",
                        "format": "float",
                        "example": 5000
                    },
                    "remaining_amount": {
                        "type": "number",
                        "format": "float",
                        "example": 5000
                    },
                    "progress_percentage": {
                        "type": "number",
                        "format": "float",
                        "example": 50
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "is_priority": {
                        "type": "boolean",
                        "example": false
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "http://example.com/image1.jpg",
                            "http://example.com/image2.jpg"
                        ]
                    },
                    "creator": {
                        "$ref": "#/components/schemas/UserResource"
                    },
                    "donations_count": {
                        "type": "integer",
                        "example": 25
                    },
                    "is_favorited": {
                        "type": "boolean",
                        "example": false
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00Z"
                    }
                },
                "type": "object"
            },
...
            "UserResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "أحمد محمد"
                    },
                    "mobile": {
                        "type": "string",
                        "example": "01234567890"
                    },
                    "wallet_balance": {
                        "type": "number",
                        "format": "float",
                        "example": 100.5
                    },
                    "profile_image": {
                        "type": "string",
                        "example": "http://example.com/image.jpg",
                        "nullable": true
                    },
                    "is_admin": {
                        "type": "boolean",
                        "example": false
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}